misc(refactor): extract config files resolver to a helper file
